requirements.txt
asgiref==3.7.2
async-generator==1.10

beautifulsoup4==4.11.1
certifi==2021.10.8
cffi==1.15.1
idna==3.3
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.0
django-heroku==0.3.1
django-mongodb-engine==0.6.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
djangotoolbox==1.8.0
dnspython==2.3.0
joblib==1.2.0
mysql-connector-python==8.0.32
numpy==1.23.5
openpyxl==3.0.10
packaging==22.0
pandas==1.3.4
pathy==0.10.1
Pillow==9.4.0
psycopg2==2.9.5
pycparser==2.21
pydantic==1.10.7
PyJWT==1.7.1
pymongo==3.12.3
PySocks==1.7.1
pytesseract==0.3.10
python-certifi-win32==1.6.1
python-dateutil==2.8.2
pytz==2021.3
regex==2023.3.23
requests==2.26.0
rest-framework-simplejwt==0.0.2
gunicorn==19.7.1
whitenoise==5.3.0
django-ckeditor==6.7.0